+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog for Vacation Soup for VR Owners
 
-## 0929.94 (2016-09-29)
+## 1007.86 (2017-10-07)
+
+* **Enhancements**:
+  - Moved Kitchen endpoint to core.vacationsoup.com
+  - Implemented message of the day (MOTD) served from core
+  - Implemented Trending Topics served from core
+  - 
+
+
+## 1002.90 (2017-10-02)
+
+Test release - not for consumption
+
+## 0929.94 (2017-09-29)
 
 Initial release for UI testing